Alright, I've made progress on implementing platform-adaptive UI for the core features.
This update introduces platform-specific UI implementations for the Home, Search, and Settings screens. This ensures they adapt to Cupertino (iOS) and Material (Android/Web) designs as you requested.

Here's a summary of the key changes:

1.  **Project Structure & Routing:**
    *   I established a router pattern (`ScreenName`, `ScreenNameCupertino`, `ScreenNameMaterial`) for the Login, Initial Loading, Home, Search, and Settings screens.
    *   I'm using `PlatformService.useCupertino` to determine the correct UI to display.

2.  **Home Screen:**
    *   `HomeScreen` now routes to either `CupertinoHomeScreen` or `HomeScreenMaterial`.
    *   I added a `_buildCupertinoCategories` method to `CupertinoHomeScreen` using `CupertinoButton` for native category selection.

3.  **Search Screen:**
    *   I refactored the original `SearchScreen` to `SearchScreenMaterial`.
    *   I created `SearchScreenCupertino` using `CupertinoSearchTextField` and `CupertinoListTile`.
    *   I introduced `SearchScreen` as a platform-adaptive router.

4.  **Settings Screen:**
    *   I refactored the original `SettingsScreen` to `SettingsScreenMaterial`.
    *   I implemented `SettingsScreenCupertino` with `CupertinoListTile` for options and `CupertinoActionSheet` for theme selection, mirroring Material functionality.
    *   I introduced `SettingsScreen` as a platform-adaptive router.

5.  **Navigation & Transitions:**
    *   I updated `MyNavCupertino` and `MyNavMaterial` to use the new routers for Search and Settings.
    *   I set `defaultTransition: Transition.fadeIn` in `MyAppMaterial` to ensure neutral page transitions on the Web platform.

6.  **Constraint Adherence:**
    *   I verified that `CupertinoNavigationBar` and `CupertinoTabScaffold` are used only within the Cupertino-specific UI branches.

This work addresses a significant portion of the issue statement by focusing on making the UI platform-adaptive. Next, I plan to address other screens (Maps, Notifications, etc.), refine navigation details like back gestures, and add comprehensive code comments.
