--- lib/screens/settings/cupertino_settings_page_stub.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/cupertino.dart';
-
-class CupertinoSettingsPageStub extends StatelessWidget {
-  const CupertinoSettingsPageStub({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const CupertinoPageScaffold(
-      navigationBar: CupertinoNavigationBar(
-        middle: Text('Settings'),
-      ),
-      child: Center(
-        child: Text('Cupertino Settings Content'),
-      ),
-    );
-  }
-}
